+++ CHANGELOG.md
@@ -6,6 +6,7 @@ versioning is done in a continuous fashion without worries of breaking changes.
 
 ## patches
 
+- `github-runners`: self host runner workflow deterministic package 2026-04-26
 - `cloud`: find stable identified protocol afront a changed proxied 2026-04-25
 - `cloud`: map secure shell addresses forbidden for cloud providing 2026-04-25
 - `github-runners`: add curl to slack github action runner packages 2026-04-25